Rename computation to coroutine
