+++ tests/coroutine.test.ts
@@ -1,230 +1,145 @@
-import { Coroutine } from "../src/coroutine";
-
-import { Call, Future, Invoke } from "../src/context";
+import { lfc, lfi, rfc, rfi } from "../src/context";
+import { Coroutine, type Suspended } from "../src/coroutine";
+import { Handler } from "../src/handler";
 
 describe("Coroutine", () => {
-  it("returns internal.return when generator is done", () => {
-    function* foo(): Generator<Invoke<any> | Future<any>, any, number> {
+  // Helper functions to write test easily
+  const exec = (uuid: string, func: any, args: any[], handler: Handler) => {
+    return new Promise<any>((resolve) => {
+      Coroutine.exec(uuid, func, args, handler, resolve);
+    });
+  };
+
+  const resolvePromise = (handler: Handler, uuid: string, value: any) => {
+    return new Promise<any>((resolve) => {
+      handler.resolvePromise(uuid, value, resolve);
+    });
+  };
+
+  test('basic coroutine completes with { type: "completed", value: 42 }', async () => {
+    function* bar() {
       return 42;
     }
 
-    const d = new Coroutine("abc", foo());
-    const r = d.next({ type: "internal.nothing", uuid: "abc" });
-
-    expect(r).toMatchObject({
-      type: "internal.return",
-      value: {
-        type: "internal.literal",
-        value: 42,
-      },
-    });
+    function* foo() {
+      const p = yield* lfi(bar);
+      const v = yield* p;
+      return v;
+    }
+    const h = new Handler();
+    const r = await exec("foo.1", foo, [], h);
+    expect(r).toEqual({ type: "completed", value: 42 });
   });
 
-  it("handles internal.async correctly", () => {
-    function* foo(): Generator<Invoke<any>, any, any> {
-      yield new Invoke("lfi", () => 42);
+  test("basic coroutine with function suspends after first await", async () => {
+    function bar() {
+      return 42;
+    }
+    function baz() {
+      return 31416;
     }
 
-    const d = new Coroutine("abc", foo());
-    const r = d.next({ type: "internal.nothing", uuid: "abc" });
+    function* foo() {
+      const p = yield* lfi(bar);
+      const p2 = yield* lfi(baz);
+      const v = yield* p;
+      const v2 = yield* p2;
+      return v + v2;
+    }
+    const h = new Handler();
 
-    expect(r).toMatchObject({
-      type: "internal.async",
-      kind: "lfi",
-      uuid: "abc.0",
-    });
+    // First execution - should suspend
+    let r = await exec("foo.1", foo, [], h);
+    expect(r).toMatchObject({ type: "suspended" });
+    const suspended = r as Suspended;
+    expect(suspended.todos).toHaveLength(2);
+
+    await resolvePromise(h, "foo.1.0", 42);
+    await resolvePromise(h, "foo.1.1", 31416);
+
+    // Second execution - should complete
+    r = await exec("foo.1", foo, [], h);
+    expect(r).toMatchObject({ type: "completed", value: 42 + 31416 });
   });
 
-  it("handles internal.await correctly", () => {
-    function* foo(): Generator<Future<any>, any, any> {
-      yield new Future("future-1", "pending");
+  test("coroutine with a suspension point suspends if can not make more progress", async () => {
+    function* bar() {
+      return 42;
     }
 
-    const d = new Coroutine("abc", foo());
-    const r = d.next({ type: "internal.nothing", uuid: "abc" });
+    function* foo() {
+      const p1 = yield* lfi(bar);
+      const p2 = yield* rfi(bar);
+      const v1 = yield* p1;
+      const vx = yield* p1;
+      const v2 = yield* p2;
+      return v1;
+    }
 
-    expect(r).toMatchObject({
-      type: "internal.await",
-      promise: {
-        uuid: "future-1",
-        state: "pending",
-      },
-    });
+    const h = new Handler();
+
+    let r = await exec("foo.1", foo, [], h);
+    expect(r.type).toBe("suspended");
+    r = r as Suspended;
+    expect(r.todos).toHaveLength(1);
+
+    await resolvePromise(h, "foo.1.1", 42);
+    r = await exec("foo.1", foo, [], h);
+    expect(r).toEqual({ type: "completed", value: 42 });
   });
 
-  it("handles lfc/rfc correctly", () => {
-    function* foo(): Generator<any, any, any> {
-      let v1 = yield new Call("lfc", (x: number) => x + 1, 1);
-      let v2 = yield new Call("rfc", (x: number) => x, 2);
+  test("Structured concurrency", async () => {
+    function* bar() {
+      return 42;
+    }
+
+    function* foo() {
+      yield* rfi(bar);
+      yield* rfi(bar);
+      return 99;
+    }
+
+    const h = new Handler();
+    let r = await exec("foo.1", foo, [], h);
+
+    expect(r.type).toBe("suspended");
+    r = r as Suspended;
+    expect(r.todos).toHaveLength(2);
+
+    await resolvePromise(h, "foo.1.1", 42);
+    r = await exec("foo.1", foo, [], h);
+
+    expect(r.type).toBe("suspended");
+    r = r as Suspended;
+    expect(r.todos).toHaveLength(1);
+
+    await resolvePromise(h, "foo.1.0", 42);
+    r = await exec("foo.1", foo, [], h);
+
+    expect(r).toEqual({ type: "completed", value: 99 });
+  });
+
+  test("lfc/rfc", async () => {
+    function* bar() {
+      return 42;
+    }
+
+    function* foo() {
+      const v1: number = yield* lfc(bar);
+      const v2: number = yield* rfc(bar);
       return v1 + v2;
     }
 
-    const d = new Coroutine("abc", foo());
-    let r = d.next({ type: "internal.nothing", uuid: "abc" });
-    expect(r).toMatchObject({
-      type: "internal.async",
-      kind: "lfi",
-    });
+    const h = new Handler();
 
-    r = d.next({
-      type: "internal.promise",
-      state: "completed",
-      uuid: "abc",
-      value: {
-        type: "internal.literal",
-        uuid: "abc",
-        value: 2,
-      },
-    });
+    let r = await exec("foo.1", foo, [], h);
+    expect(r.type).toBe("suspended");
+    let suspended = r as Suspended;
+    expect(suspended.todos).toHaveLength(1);
 
-    expect(r).toMatchObject({
-      type: "internal.await",
-      promise: {
-        type: "internal.promise",
-        state: "completed",
-        value: {
-          type: "internal.literal",
-          value: 2,
-        },
-      },
-    });
+    await resolvePromise(h, "foo.1.1", 42);
 
-    r = d.next({
-      type: "internal.literal",
-      value: 2,
-      uuid: "abc",
-    });
-
-    expect(r).toMatchObject({
-      type: "internal.async",
-      kind: "rfi",
-    });
-
-    r = d.next({
-      type: "internal.promise",
-      state: "pending",
-      uuid: "abc",
-    });
-
-    expect(r).toMatchObject({
-      type: "internal.await",
-      promise: {
-        type: "internal.promise",
-        state: "pending",
-      },
-    });
-  });
-
-  it("returns final value after multiple yields", () => {
-    function* foo(): Generator<Invoke<any> | Future<any>, any, number> {
-      yield new Future("future-1", "completed", 10);
-      yield new Invoke("lfi", () => 42);
-      return 30;
-    }
-
-    const d = new Coroutine("abc", foo());
-
-    d.next({ type: "internal.nothing", uuid: "abc" }); // First yield
-    d.next({ type: "internal.literal", uuid: "abc.1", value: 10 }); // yield a future, get a literal back
-    const r = d.next({
-      type: "internal.promise",
-      state: "completed",
-      uuid: "abc.2",
-      value: {
-        type: "internal.literal",
-        uuid: "abc.2.lit",
-        value: 42,
-      },
-    }); // yield an invoke, get a completed promise back
-
-    expect(r).toMatchObject({
-      type: "internal.return",
-      value: {
-        type: "internal.literal",
-        value: 30,
-      },
-    });
-  });
-
-  it("awaits if there are pending invokes - Structured Concurrency", () => {
-    function* foo(): Generator<Invoke<any> | Future<any>, any, number> {
-      yield new Future("future-1", "completed", 10); // A
-      yield new Invoke("lfi", () => 20); // B
-      yield new Invoke("lfi", () => 30); // C
-      return 30; // D
-    }
-
-    const d = new Coroutine("abc", foo());
-
-    d.next({ type: "internal.nothing", uuid: "abc" }); // First yield
-    d.next({ type: "internal.literal", uuid: "abc.1", value: 10 }); // A -> yield a future, get a literal back
-    d.next({
-      type: "internal.promise",
-      state: "pending",
-      uuid: "abc.1",
-    }); // B -> yield an invoke, get a pending promise back
-    const r = d.next({
-      type: "internal.promise",
-      state: "completed",
-      uuid: "abc.2",
-      value: {
-        type: "internal.literal",
-        uuid: "abc.1.lit",
-        value: 30,
-      },
-    }); // C -> yield an invoke, get a completed promise back
-
-    // D -> Must not return, instead tell the caller to await given that there is a pending invoke
-    expect(r).toMatchObject({
-      type: "internal.await",
-      promise: {
-        type: "internal.promise",
-        state: "pending",
-      },
-    });
-  });
-
-  it("returns if there are no pending invokes even if not explecityly awaited", () => {
-    function* foo(): Generator<Invoke<any> | Future<any>, any, number> {
-      yield new Future("future-1", "completed", 10); // A
-      yield new Invoke("lfi", () => 20); // B
-      yield new Invoke("lfi", () => 30); // C
-      return 42; // D
-    }
-
-    const d = new Coroutine("abc", foo());
-
-    d.next({ type: "internal.nothing", uuid: "abc" }); // First yield
-    d.next({ type: "internal.literal", uuid: "abc.1", value: 10 }); // A -> yield a future, get a literal back
-    d.next({
-      type: "internal.promise",
-      state: "completed",
-      uuid: "abc.1",
-      value: {
-        type: "internal.literal",
-        uuid: "abc.1.lit",
-        value: 20,
-      },
-    }); // B -> yield an invoke, get a completed promise back
-    const r = d.next({
-      type: "internal.promise",
-      state: "completed",
-      uuid: "abc.2",
-      value: {
-        type: "internal.literal",
-        uuid: "abc.1.lit",
-        value: 30,
-      },
-    }); // C -> yield an invoke, get a completed promise back
-
-    // D -> Must return given that the previous invokes were completed
-    // even if they were not explicitly awaited.
-    expect(r).toMatchObject({
-      type: "internal.return",
-      value: {
-        type: "internal.literal",
-        value: 42,
-      },
-    });
+    r = await exec("foo.1", foo, [], h);
+    expect(r).toEqual({ type: "completed", value: 84 });
   });
 });
